+++ common/src/serialization/primitive.cpp
@@ -1,4 +1,4 @@
-#include "primitive.h"
+#include "serialization/primitive.h"
 
 using namespace serialization;
