+++ common/src/serialization/request.cpp
@@ -1,5 +1,5 @@
-#include "request.h"
-#include "size_collector.h"
+#include <serialization/request.h>
+#include <serialization/size_collector.h>
 
 using namespace serialization;
