+++ common/src/serialization/response.cpp
@@ -1,5 +1,5 @@
-#include "response.h"
-#include "size_collector.h"
+#include "serialization/response.h"
+#include "serialization/size_collector.h"
 
 using namespace serialization;
