+++ server/include/Server.h
@@ -4,7 +4,7 @@
 #include <tbb/concurrent_unordered_map.h>
 #include <model/request.h>
 #include <model/response.h>
-#include "lib/ctpl.h"
+#include <lib/ctpl.h>
 #include "EmailWithInfo.h"
 
 class Server {
