+++ server/src/main.cpp
@@ -1,4 +1,4 @@
-#include "Server.h"
+#include "../include/Server.h"
 
 #define PORT 4567
